requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.84.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
